Get proper nested scroll/event handling behavior with drei Html
By attaching the HTML to the appropriate parent (in my case, it seems to be the scroll controller's "fixed" element), it looks like the nested event handling behavior Just Works... or at least based on my limited testing. I'm worried that there's some kind of catch to this though that makes it too good to be true.