Change how cable progress is mapped
It doesn't really have to do with mobile improvements, but I wanted to do it.

Basically, I think there are 2 kinds of mappings along the length of the curve - 1 for like a progress from point to point that defines the original curve, and then another that factors in the length of the entire curve, and then subdivides the curve into equal-length subdivisions and then calculates a progress based on that value. I think I was using the former before (which would give inconsistent speeds depending on which part of the curve I was on - it was especially noticeable on the sharper turns where it would go really quickly), but the latter gives a smoother result, I think