Switch to phone model and start to update display
After learning more about how HTML works in Three, I'm gonna have to switch my tactic a bit to hide the fact that I basically can't get proper masking/occlusion with the model working as intended to get round corners. So, instead, I'm gonna use non-rounded corners and try to blend it in to the black screen background thing to fake it.