Try to reduce text flicker on Chrome
I'm not totally sure what the issue is, but I suspect it is either related to performance (in which case, hopefully the previous commit helped) or because of scaling the text and floating point precision. This commit is mostly an attempt to help with the latter, though I also tweaked some timing to hopefully help with potential performance issues as well.