Improve scrolling issue with link page
For some reason that I still haven't totally figured out, on mobile I'm not able to scroll the entire way down the page, which is allowing the Links page to get cut off. I'm guessing this broke during the switch to Lenis, but I missed it until recently.

For now, use the "syncTouch" option along with making the screen height slightly more dynamic to improve things. There's now an issue though where I can't seem to get the search/URL bar to disappear as expected on mobile, but it's still a better experience than the content getting cut off, so going with it for now.