Get parallax working on every frame
Before, there was an issue where the parallax would only update on scroll. However, this meant that the initial position might be wrong until the user scrolls, causing a visual jump.

Also, this helps handle cases where positions might change for reasons other than scroll.

It may not be the most optimal, but it seems to work well enough for now.